Moved ClarinUtils into Utils class
