+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/security/clarin/ClarinShibbolethLoginFilter.java
@@ -22,7 +22,6 @@ import org.apache.logging.log4j.Logger;
 import org.dspace.app.rest.security.DSpaceAuthentication;
 import org.dspace.app.rest.security.RestAuthenticationService;
 import org.dspace.app.rest.security.StatelessLoginFilter;
-import org.dspace.app.rest.utils.ClarinUtils;
 import org.dspace.authenticate.clarin.ClarinShibAuthentication;
 import org.dspace.authenticate.clarin.ShibHeaders;
 import org.dspace.content.clarin.ClarinVerificationToken;
@@ -309,7 +308,7 @@ public class ClarinShibbolethLoginFilter extends StatelessLoginFilter {
         if (StringUtils.equalsAnyIgnoreCase(redirectHostName, allowedHostNames.toArray(new String[0]))) {
             log.debug("Shibboleth redirecting to " + redirectUrl);
             // Encode the UTF-8 characters from redirect URL to UTF-8, to ensure it's properly encoded for the browser
-            String encodedRedirectUrl = ClarinUtils.encodeNonAsciiCharacters(redirectUrl);
+            String encodedRedirectUrl = org.dspace.app.rest.dq.Utils.encodeNonAsciiCharacters(redirectUrl);
             if (StringUtils.isEmpty(encodedRedirectUrl)) {
                 log.error("Invalid Encoded Shibboleth redirectURL=" + redirectUrl + ". URL is empty!");
             }
